pyproject.toml
[project]
name = "warehouse-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]
